Create new accounts.toml config file to store addresses
