+++ cmd/wallet_migrate.go
@@ -29,7 +29,7 @@ var migrateCmd = &cobra.Command{
 
 		fmt.Printf("Please set a new passphrase to encrypt the owner key:\n\n")
 
-		as := util.AgentStore()
+		as := util.AccountsStore()
 		ownerAddr, _, err := as.GetAddrs(util.OwnerKey)
 		if err != nil {
 			logFatal(err)
@@ -61,7 +61,7 @@ func migrateLegacyKeys() error {
 func migrateLegacyKey(key util.KeyType) error {
 	ksLegacy := util.KeyStoreLegacy()
 	ks := util.KeyStore()
-	as := util.AgentStore()
+	as := util.AccountsStore()
 
 	pkStr, err := ksLegacy.Get(string(key))
 	if err != nil {
